backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.5.1
python-dotenv==1.0.1
eventlet==0.33.3
pymongo==4.10.1
gunicorn==20.1.0
Flask-Bcrypt==1.0.1
Flask-Bcrypt==1.0.1    # For hashing passwords
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.5.1
Flask==3.0.3
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
eventlet==0.33.3
flask-cors==5.0.0
flask-socketio==5.5.1
gunicorn==20.1.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
pymongo==4.10.1
python-dotenv==1.0.1
zipp==3.21.0
